+++ lib/view/community/content.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:han_bab/view/community/plus_page.dart';
+import 'package:han_bab/view/community/pluspage.dart';
 import 'package:like_button/like_button.dart';
 
 import '../main/community_page.dart';
