+++ lib/view/main/main_screen.dart
@@ -4,7 +4,7 @@ import 'package:han_bab/controller/main_bottom_bar_controller.dart';
 import 'package:han_bab/model/main_bottom_bar.dart';
 import 'package:han_bab/view/main/profile_page.dart';
 import 'package:han_bab/view/main/home_page.dart';
-import 'package:han_bab/view/main/my_page.dart';
+import 'package:han_bab/view/main/profile_page.dart';
 import '../../component/stateful_wrapper.dart';
 import '../main/home_page.dart';
 import 'community_page.dart';
