--- lib/view/mypage/profile_page.dart
@@ -1,2 +1,0 @@
-import 'package:flutter/material.dart';
-
